fix: seperate concerns for extension alerts
